requirements.txt
python-telegram-bot==22.6
websockets>=13.0,<14
python-dotenv>=1.0
